Website set up
I have changed coding for the layout of the website so that everything is alligned within the margins and the screen should change according to size.

I have also added images to the home screen and created links for them.

The basics have been inputted for the contact and about page for the theme and my section on the about page.

I have attempted to add a contact form on the page and have changed some of the images that were already saved onto the group.